Split SmartWalletService into focused single-responsibility service protocols
