+++ Sources/Wallet/Data/Services/BalanceService.swift
@@ -1,0 +1,8 @@
+import CrossmintCommonTypes
+import CrossmintService
+
+public protocol BalanceService: AuthenticatedService, Sendable {
+    func getBalance(
+        _ params: GetBalanceQueryParams
+    ) async throws(WalletError) -> Balances
+}
