+++ Sources/Wallet/Data/Services/NFTService.swift
@@ -1,0 +1,8 @@
+import CrossmintCommonTypes
+import CrossmintService
+
+public protocol NFTService: AuthenticatedService, Sendable {
+    func getNFTs(
+        _ params: GetNTFQueryParams
+    ) async throws(WalletError) -> [NFT]
+}
