+++ Sources/Wallet/Data/Services/TransactionService.swift
@@ -1,0 +1,15 @@
+import CrossmintService
+
+public protocol TransactionService: AuthenticatedService, Sendable {
+    func createTransaction(
+        _ request: CreateTransactionRequest
+    ) async throws(TransactionError) -> any TransactionApiModel
+
+    func signTransaction(
+        _ request: SignRequest
+    ) async throws(TransactionError) -> any TransactionApiModel
+
+    func fetchTransaction(
+        _ request: FetchTransactionRequest
+    ) async throws(TransactionError) -> any TransactionApiModel
+}
